scripts/requirements.txt
astroid==2.3.3
autopep8==1.4.4
bcrypt==3.1.7
boto3==1.10.46
botocore==1.13.46
certifi==2019.11.28
cffi==1.13.2
cryptography==2.8
DateTime==4.3
docutils==0.15.2
fabric==2.5.0
invoke==1.3.0
isort==4.3.21
jmespath==0.9.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
paramiko==2.7.1
pycodestyle==2.5.0
pycparser==2.19
pylint==2.4.4
PyNaCl==1.3.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
s3transfer==0.2.1
six==1.13.0
typed-ast==1.4.1
urllib3==1.25.7
wrapt==1.11.2
zope.interface==4.7.1
